requirements.txt
absl-py==2.1.0
amqp==5.2.0
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==41.0.7
daphne==4.0.0
defusedxml==0.7.1
Deprecated==1.2.14
diff-match-patch==20230430
Django==4.2.8
django-ajax-selects==3.0.0
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-filter==23.5
django-import-export==3.3.4
django-phonenumber-field==7.2.0
django-rest-knox==4.2.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
dm-tree==0.1.8
et-xmlfile==1.1.0
flatbuffers==23.5.26
gast==0.5.4
gevent==23.9.1
google-pasta==0.2.0
greenlet==3.0.0rc3
grpcio==1.62.0
gunicorn==21.2.0
h5py==3.10.0
hyperlink==21.0.0
idna==3.6
importlib-metadata==7.0.0
incremental==22.10.0
keras==3.0.5
kombu==5.3.4
libclang==16.0.6
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupPy==1.14
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
msgpack==1.0.7
msgpack-python==0.5.6
namex==0.0.7
numpy==1.26.2
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.2
phonenumbers==8.13.26
Pillow==10.1.0
prompt-toolkit==3.0.41
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
service-identity==23.1.0
setuptools==69.1.0
six==1.16.0
sorl-thumbnail==12.10.0
sqlparse==0.4.4
tablib==3.5.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow-intel==2.16.0rc0
termcolor==2.4.0
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
wheel==0.42.0
wrapt==1.16.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
